Refactor Accounts.tsx into /account/index.tsx
